move PDE related class into pde.hpp
