+++ examples/finitevolume.cpp
@@ -28,7 +28,7 @@
 #include <mpi.h>
 
 #include "mfem.hpp"
-#include "spe10.hpp"
+#include "pde.hpp"
 
 #include "../src/picojson.h"
 #include "../src/smoothG.hpp"
