+++ examples/mlmc.cpp
@@ -32,8 +32,7 @@
 #include <sstream>
 #include <mpi.h>
 
-#include "mfem.hpp"
-#include "spe10.hpp"
+#include "pde.hpp"
 
 #include "../src/picojson.h"
 #include "../src/smoothG.hpp"
