+++ examples/timestep.cpp
@@ -22,8 +22,7 @@
 #include <sstream>
 #include <mpi.h>
 
-#include "mfem.hpp"
-#include "spe10.hpp"
+#include "pde.hpp"
 
 #include "../src/picojson.h"
 #include "../src/smoothG.hpp"
