+++ src/utilities.cpp
@@ -294,47 +294,6 @@ void VisualizeSolution(int k,
     Dsigma.Save(Dsigma_ofs);
 }
 
-/**
-   Construct edge to boundary attribute table (orientation is not considered)
-*/
-mfem::SparseMatrix GenerateBoundaryAttributeTable(const mfem::Mesh* mesh)
-{
-    int nedges = mesh->Dimension() == 2 ? mesh->GetNEdges() : mesh->GetNFaces();
-    int nbdr = mesh->bdr_attributes.Max();
-    int nbdr_edges = mesh->GetNBE();
-
-    int* edge_bdrattr_i = new int[nedges + 1]();
-    int* edge_bdrattr_j = new int[nbdr_edges];
-
-    // in the loop below, edge_bdrattr_i is used as a temporary array
-    for (int j = 0; j < nbdr_edges; j++)
-    {
-        int edge = mesh->GetBdrElementEdgeIndex(j);
-        edge_bdrattr_i[edge + 1] = mesh->GetBdrAttribute(j);
-    }
-    edge_bdrattr_i[0] = 0;
-
-    int count = 0;
-    for (int j = 1; j <= nedges; j++)
-    {
-        if (edge_bdrattr_i[j])
-        {
-            edge_bdrattr_j[count++] = edge_bdrattr_i[j] - 1;
-            edge_bdrattr_i[j] = edge_bdrattr_i[j - 1] + 1; // single nonzero in this row
-        }
-        else
-        {
-            edge_bdrattr_i[j] = edge_bdrattr_i[j - 1]; // no nonzeros in this row
-        }
-    }
-
-    double* edge_bdrattr_data = new double[nbdr_edges];
-    std::fill_n(edge_bdrattr_data, nbdr_edges, 1.0);
-
-    return mfem::SparseMatrix(edge_bdrattr_i, edge_bdrattr_j, edge_bdrattr_data,
-                              nedges, nbdr);
-}
-
 int MarkDofsOnBoundary(
     const mfem::SparseMatrix& face_boundaryatt,
     const mfem::SparseMatrix& face_dof,
@@ -575,171 +534,6 @@ void ReadCoordinate(std::ifstream& graphFile, mfem::SparseMatrix& out)
 
     out.Swap(mat);
 }
-
-void InversePermeabilityFunction::SetNumberCells(int Nx_, int Ny_, int Nz_)
-{
-    Nx = Nx_;
-    Ny = Ny_;
-    Nz = Nz_;
-}
-
-void InversePermeabilityFunction::SetMeshSizes(double hx_, double hy_,
-                                               double hz_)
-{
-    hx = hx_;
-    hy = hy_;
-    hz = hz_;
-}
-
-void InversePermeabilityFunction::Set2DSlice(SliceOrientation o, int npos_ )
-{
-    orientation = o;
-    npos = npos_;
-}
-
-void InversePermeabilityFunction::BlankPermeability()
-{
-    inversePermeability = new double[3 * Nx * Ny * Nz];
-    for (int i = 0; i < 3 * Nx * Ny * Nz; ++i)
-    {
-        inversePermeability[i] = 1.0;
-    }
-}
-
-void InversePermeabilityFunction::ReadPermeabilityFile(const std::string& fileName)
-{
-    std::ifstream permfile(fileName.c_str());
-
-    if (!permfile.is_open())
-    {
-        std::cerr << "Error in opening file " << fileName << std::endl;
-        mfem::mfem_error("File does not exist");
-    }
-
-    inversePermeability = new double [3 * Nx * Ny * Nz];
-    double* ip = inversePermeability;
-    double tmp;
-    for (int l = 0; l < 3; l++)
-    {
-        for (int k = 0; k < Nz; k++)
-        {
-            for (int j = 0; j < Ny; j++)
-            {
-                for (int i = 0; i < Nx; i++)
-                {
-                    permfile >> *ip;
-                    *ip = 1. / (*ip);
-                    ip++;
-                }
-                for (int i = 0; i < 60 - Nx; i++)
-                    permfile >> tmp; // skip unneeded part
-            }
-            for (int j = 0; j < 220 - Ny; j++)
-                for (int i = 0; i < 60; i++)
-                    permfile >> tmp;  // skip unneeded part
-        }
-
-        if (l < 2) // if not processing Kz, skip unneeded part
-            for (int k = 0; k < 85 - Nz; k++)
-                for (int j = 0; j < 220; j++)
-                    for (int i = 0; i < 60; i++)
-                        permfile >> tmp;
-    }
-
-}
-
-void InversePermeabilityFunction::ReadPermeabilityFile(const std::string& fileName,
-                                                       MPI_Comm comm)
-{
-    int num_procs, myid;
-    MPI_Comm_size(comm, &num_procs);
-    MPI_Comm_rank(comm, &myid);
-
-    mfem::StopWatch chrono;
-
-    chrono.Start();
-    if (myid == 0)
-        ReadPermeabilityFile(fileName);
-    else
-        inversePermeability = new double [3 * Nx * Ny * Nz];
-    chrono.Stop();
-
-    if (myid == 0)
-        std::cout << "Permeability file read in " << chrono.RealTime() << ".s \n";
-
-    chrono.Clear();
-
-    chrono.Start();
-    MPI_Bcast(inversePermeability, 3 * Nx * Ny * Nz, MPI_DOUBLE, 0, comm);
-    chrono.Stop();
-
-    if (myid == 0)
-        std::cout << "Permeability field distributed in " << chrono.RealTime() << ".s \n";
-
-}
-
-void InversePermeabilityFunction::InversePermeability(const mfem::Vector& x,
-                                                      mfem::Vector& val)
-{
-    val.SetSize(x.Size());
-
-    unsigned int i = 0, j = 0, k = 0;
-
-    switch (orientation)
-    {
-        case NONE:
-            i = Nx - 1 - (int)floor(x[0] / hx / (1. + 3e-16));
-            j = (int)floor(x[1] / hy / (1. + 3e-16));
-            k = Nz - 1 - (int)floor(x[2] / hz / (1. + 3e-16));
-            break;
-        case XY:
-            i = Nx - 1 - (int)floor(x[0] / hx / (1. + 3e-16));
-            j = (int)floor(x[1] / hy / (1. + 3e-16));
-            k = npos;
-            break;
-        case XZ:
-            i = Nx - 1 - (int)floor(x[0] / hx / (1. + 3e-16));
-            j = npos;
-            k = Nz - 1 - (int)floor(x[2] / hz / (1. + 3e-16));
-            break;
-        case YZ:
-            i = npos;
-            j = (int)floor(x[1] / hy / (1. + 3e-16));
-            k = Nz - 1 - (int)floor(x[2] / hz / (1. + 3e-16));
-            break;
-        default:
-            mfem::mfem_error("InversePermeabilityFunction::InversePermeability");
-    }
-
-    val[0] = inversePermeability[Ny * Nx * k + Nx * j + i];
-    val[1] = inversePermeability[Ny * Nx * k + Nx * j + i + Nx * Ny * Nz];
-
-    if (orientation == NONE)
-        val[2] = inversePermeability[Ny * Nx * k + Nx * j + i + 2 * Nx * Ny * Nz];
-}
-
-double InversePermeabilityFunction::InvNorm2(const mfem::Vector& x)
-{
-    mfem::Vector val(3);
-    InversePermeability(x, val);
-    return 1. / val.Norml2();
-}
-
-void InversePermeabilityFunction::ClearMemory()
-{
-    delete[] inversePermeability;
-}
-
-int InversePermeabilityFunction::Nx(60);
-int InversePermeabilityFunction::Ny(220);
-int InversePermeabilityFunction::Nz(85);
-double InversePermeabilityFunction::hx(20);
-double InversePermeabilityFunction::hy(10);
-double InversePermeabilityFunction::hz(2);
-double* InversePermeabilityFunction::inversePermeability(NULL);
-InversePermeabilityFunction::SliceOrientation InversePermeabilityFunction::orientation(
-    InversePermeabilityFunction::NONE );
-int InversePermeabilityFunction::npos(-1);
 
 double DivError(MPI_Comm comm, const mfem::SparseMatrix& D, const mfem::Vector& numer,
                 const mfem::Vector& denom)
